requirements.txt
celery==4.2.1
grpcio==1.19.0
grpcio-tools==1.19.0
ontology-python-sdk==1.3.1
protobuf==3.7.0
Flask==1.0.2
gunicorn==19.9.0